Move _PullToPush to common file.
