+++ src/twisted/web/http.py
@@ -98,8 +98,8 @@ from twisted.python.components import proxyForInterface
 from twisted.internet import interfaces, protocol, address
 from twisted.internet.defer import Deferred
 from twisted.internet.interfaces import IProtocol
+from twisted.internet.producer_helpers import _PullToPush
 from twisted.protocols import policies, basic
-from twisted.protocols.tls import _PullToPush
 
 from twisted.web.iweb import (
     IRequest, IAccessLogFormatter, INonQueuedRequestFactory)
